+++ src/worker/02-parseWiki.js
@@ -12,16 +12,21 @@ const plaintext = function(page) {
 
 //get parsed json from the wiki markup
 const parseData = function(page, options) {
-  let data = wtf.parse(page.script);
+  let doc = wtf(page.script);
   //dont insert this if it's a redirect
-  if (options.skip_redirects === true && data.type === 'redirect') {
+  if (options.skip_redirects === true && doc.isRedirect()) {
     return null
   }
-  if (options.skip_disambig === true && data.type === 'disambiguation') {
+  if (options.skip_disambig === true && doc.isDisambiguation()) {
     return null
   }
+  //turn the wtf_wikipedia document into storable json
+  let data = doc.json(options)
+  data.title = page.title || data.title
   data = encode.encodeData(data);
-  data.title = data.title || page.title || '';
+  //use the title/pageID from the xml
+  data.title = page.title || data.title || '';
+  data.pageID = page.pageID || data.pageID || '';
   data._id = encode.encodeStr(data.title);
   return data
 };