+++ changelog.md
@@ -10,3 +10,7 @@
 ### v2.4.0
 * add a 3s 'break' to avoid build-up of mongo inserts
 * add new --verbose and --skip_first options
+
+## v3
+* MASSIVE SPEEDUP! full re-write by @devrim 🙏 to fix #59
+* use wtf_wikipedia v3 (big re-factor too!)